+++ CHANGELOG.md
@@ -4,10 +4,13 @@ All notable changes to the Wazuh app project will be documented in this file.
 
 ## Wazuh v4.3.6 - Kibana 7.10.2, 7.16.x, 7.17.x - Revision 4307
 
+### Added
+- Added a new documentation link to the Docker Listener module [#4301](https://github.com/wazuh/wazuh-kibana-app/pull/4301)
 ### Changed
 
 - The links to the web documentation now points to the plugin short version instead of `current` [#4301](https://github.com/wazuh/wazuh-kibana-app/pull/4301)
 - Changed some prompts that didn't match the with the current style [#4301](https://github.com/wazuh/wazuh-kibana-app/pull/4301)
+- Changed some link titles to match their documentation section title [#4301](https://github.com/wazuh/)
 
 # Fixed